requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.23
colorama==0.4.6
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
duckdb==0.7.1
entrypoints==0.4
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.95.1
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.0.150
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
posthog==3.0.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pycryptodome==3.17
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyPDF2==3.0.1
pypdf==3.8.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.12
stack-data==0.6.2
starlette==0.26.1
streamlit==1.22.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0